+++ packages/tenderly-core/src/internal/core/services/TenderlyApiService.ts
@@ -3,21 +3,31 @@ import * as axios from "axios";
 import { logger } from "../../../utils/logger";
 import { getConfig, isAccessTokenSet } from "../../../utils/config";
 import { TENDERLY_API_BASE_URL, TENDERLY_JSON_RPC_BASE_URL } from "../../../common/constants";
+import {HARDHAT_TENDERLY_VERSION} from "./hardhat-tenderly-version";
 
 export class TenderlyApiService {
-  public static configureInstance(): axios.AxiosInstance {
+  public static configureInstance(providedAccessKey?: string): axios.AxiosInstance {
     logger.debug("Configuring instance.");
 
-    const tdlyConfig = getConfig();
+    let accessKey: string
+    if (!!providedAccessKey) {
+      accessKey = providedAccessKey
+    } else {
+      const tdlyConfig = getConfig();
+      accessKey = tdlyConfig.access_key
+    }
     const params = {
       baseURL: TENDERLY_API_BASE_URL,
-      headers: { "x-access-key": tdlyConfig.access_key },
+      headers: { 
+        "x-access-key": accessKey,
+        "x-user-agent": `@tenderly/hardhat-tenderly/${HARDHAT_TENDERLY_VERSION}`
+      },
     };
 
     logger.debug("Configured instance with parameters:", {
       baseURL: TENDERLY_API_BASE_URL,
       access_key:
-        tdlyConfig.access_key !== undefined && tdlyConfig.access_key !== null && tdlyConfig.access_key !== ""
+      accessKey !== undefined && accessKey !== null && accessKey !== ""
           ? "super secret access_key is set in 'access_key' field"
           : "undefined or null or empty string",
     });